Better semantic HTML, exposing of classes for customization
Updated the HTML to output the dump legend using an HTML legend tag
with a class on it to make CSS styling easier, versus using a
hardcoded h2, which is not semantically correct.  Also provided a
few overridable class properties in other places to allow e.g.
setting the value column to pull-right versus pull-left, or any
other CSS styling desired.